+++ test/plug/csrf_protection_test.exs
@@ -6,36 +6,27 @@ defmodule Plug.CSRFProtectionTest do
   alias Plug.CSRFProtection.InvalidCSRFTokenError
   alias Plug.CSRFProtection.InvalidCrossOriginRequestError
 
-  @default_opts Plug.Session.init(
-    store: :cookie,
-    key: "foobar",
-    encryption_salt: "cookie store encryption salt",
-    signing_salt: "cookie store signing salt",
-    encrypt: true
-  )
-
-  @secret String.duplicate("abcdef0123456789", 8)
   @csrf_token "hello123"
 
-  def call(conn) do
+  def call(conn, opts \\ [])
+
+  def call(conn, opts) when is_list(opts) do
     conn
-    |> sign_cookie(@secret)
     |> fetch_params
-    |> Plug.Session.call(@default_opts)
-    |> fetch_session
-    |> CSRFProtection.call([])
+    |> fetch_cookies
+    |> CSRFProtection.call(opts)
     |> put_resp_content_type(conn.assigns[:content_type] || "text/html")
     |> send_resp(200, "ok")
   end
 
-  def call(conn, old_conn) do
-    conn
-    |> recycle_cookies(old_conn)
-    |> call()
+  def call(conn, old_conn) when is_map(old_conn) do
+    call(conn, old_conn, [])
   end
 
-  defp sign_cookie(conn, secret) do
-    put_in conn.secret_key_base, secret
+  def call(conn, old_conn, opts) do
+    conn
+    |> recycle_cookies(old_conn)
+    |> call(opts)
   end
 
   test "raise error for missing authenticity token in session" do
@@ -65,16 +56,16 @@ defmodule Plug.CSRFProtectionTest do
   test "unprotected requests are always valid" do
     conn = conn(:get, "/") |> call()
     assert conn.halted == false
-    assert get_session(conn, :csrf_token)
+    assert conn.resp_cookies["_csrf_token"]
 
     conn = conn(:head, "/") |> call()
     assert conn.halted == false
-    assert get_session(conn, :csrf_token)
+    assert conn.resp_cookies["_csrf_token"]
   end
 
   test "protected requests with valid token in params are allowed" do
     old_conn = conn(:get, "/") |> call
-    params = %{csrf_token: get_session(old_conn, :csrf_token)}
+    params = %{_csrf_token: old_conn.resp_cookies["_csrf_token"].value}
 
     conn = conn(:post, "/", params) |> call(old_conn)
     assert conn.halted == false
@@ -87,8 +78,8 @@ defmodule Plug.CSRFProtectionTest do
   end
 
   test "protected requests with valid token in header are allowed" do
-    old_conn = conn(:get, "/") |> call
-    csrf_token = get_session(old_conn, :csrf_token)
+    old_conn   = conn(:get, "/") |> call
+    csrf_token = old_conn.resp_cookies["_csrf_token"].value
 
     conn =
       conn(:post, "/")
@@ -109,6 +100,16 @@ defmodule Plug.CSRFProtectionTest do
     assert conn.halted == false
   end
 
+  test "protected requests with valid token and custom name" do
+    conn = conn(:get, "/") |> call(name: "SEKRET")
+    assert conn.halted == false
+    assert conn.resp_cookies["SEKRET"]
+
+    params = %{_csrf_token: conn.resp_cookies["SEKRET"].value}
+    conn = conn(:post, "/", params) |> call(conn, name: "SEKRET")
+    assert conn.halted == false
+  end
+
   test "non-XHR Javascript GET requests are forbidden" do
     assert_raise InvalidCrossOriginRequestError, fn ->
       conn(:get, "/") |> assign(:content_type, "application/javascript") |> call()
@@ -125,7 +126,7 @@ defmodule Plug.CSRFProtectionTest do
       |> assign(:content_type, "text/javascript")
       |> put_req_header("x-requested-with", "XMLHttpRequest")
       |> call()
-    assert get_session(conn, :csrf_token)
+    assert conn.resp_cookies["_csrf_token"]
   end
 
   test "csrf plug is skipped when plug_skip_csrf_protection is true" do
@@ -133,19 +134,19 @@ defmodule Plug.CSRFProtectionTest do
       conn(:get, "/")
       |> put_private(:plug_skip_csrf_protection, true)
       |> call()
-    assert get_session(conn, :csrf_token)
+    assert conn.resp_cookies["_csrf_token"]
 
     conn =
       conn(:post, "/", %{})
       |> put_private(:plug_skip_csrf_protection, true)
       |> call()
-    assert get_session(conn, :csrf_token)
+    assert conn.resp_cookies["_csrf_token"]
 
     conn =
       conn(:get, "/")
       |> put_private(:plug_skip_csrf_protection, true)
       |> assign(:content_type, "text/javascript")
       |> call()
-    assert get_session(conn, :csrf_token)
+    assert conn.resp_cookies["_csrf_token"]
   end
 end